# Use `__getattr__` for all package imports, improve import time