--- a/docs/api/validate_call.md
+++ b/docs/api/validate_call.md
@@ -1,1 +1,1 @@
-::: pydantic.validate_call
+::: pydantic.validate_call_decorator