… docs/concepts/strict_mode.md
… docs/concepts/validation_decorator.md
… mkdocs.yml
… pydantic/types.py
… pydantic/validate_call_decorator.py
… pydantic/version.py
… tests/test_exports.py