@ofek 2023-10-27
Awesome thank you, definitely a step in the right direction!

```
❯ pip install -Uq pydantic
❯ python -m timeit -n 1 -r 1 "import pydantic"
1 loop, best of 1: 83.4 msec per loop
❯ python -m timeit -n 1 -r 1 "from pydantic import BaseModel"
1 loop, best of 1: 89.6 msec per loop
❯ pip install -Iq "pydantic @ git+https://github.com/pydantic/pydantic.git@defer-init-imports"
❯ python -m timeit -n 1 -r 1 "import pydantic"
1 loop, best of 1: 15.7 msec per loop
❯ python -m timeit -n 1 -r 1 "from pydantic import BaseModel"
1 loop, best of 1: 79 msec per loop
```

Do you see a way to reduce the import cost of `BaseModel` any further? That is what I, and I think most others, are optimizing for because that is the main interface that classes in one's code will inherit at import time.