@ofek 2023-10-27
That last commit helped a little!

```
❯ python -m timeit -n 1 -r 1 "from pydantic import BaseModel"
1 loop, best of 1: 63.9 msec per loop
```