@ofek 2023-10-27
![image](https://github.com/pydantic/pydantic/assets/9677399/fe9ac448-d125-48ba-9241-d03272b4c351)