@samuelcolvin 2023-10-27
@ofek I've got rid of import time import of `annotated-types`, with that I see a >38% improvement. I really don't think there's much more I can do from there.

If you see anything else, let me know.