@ofek 2023-10-27
Thank you!

```
❯ python -m timeit -n 1 -r 1 "from pydantic import BaseModel"
1 loop, best of 1: 56.8 msec per loop
```

![image](https://github.com/pydantic/pydantic/assets/9677399/72b5cc6c-81a6-45e0-8378-62aa344ca4f7)