@samuelcolvin 2023-10-27
I don't think anything in logs, I think it merits a prominent notice in the release notes and a tweet from me now saying we're doing it.

AFAIK this shouldn't affect anyone and I made the change in few steps to see, but the only real answer is to release it and see.