@ofek 2023-10-27
Out of curiosity, how will the practice of being conscious about import time be maintained during continued development?