@samuelcolvin 2023-10-27
I've added unit tests that that those modules aren't imported in specific cases, e.g. `from pydantic import BaseModel`.

The longer term solution is all down to @art049 - we should use Codspeed for continuous profile of stuff we care about like import time, as we do on pydantic-core.